requirements.txt
requests
falcon
numpy
pandas
scipy
gunicorn
msgpack-python
